+++ model_navigator/commands/export/exporters/torch2dynamo_onnx.py
@@ -13,9 +13,12 @@
 # limitations under the License.
 """Export Torch model using dynamo."""
 
+import logging
 import pathlib
-from typing import Any, Dict, Optional
+from typing import Any, Dict, List, Optional
 
+import onnx
+import onnx_graphsurgeon as gs
 import torch  # pytype: disable=import-error
 
 from model_navigator.core.dataloader import load_samples
@@ -34,8 +37,12 @@ def get_model() -> torch.nn.Module:
 def export(
     exported_model_path: str,
     input_metadata: Dict[str, Any],
+    input_names: List[str],
+    output_names: List[str],
     batch_dim: Optional[int],
     target_device: str,
+    dynamic_shapes: bool,
+    verbose: bool,
     custom_args: Dict[str, Any],
     navigator_workspace: Optional[str] = None,
 ):
@@ -44,9 +51,12 @@ def export(
     Args:
         exported_model_path (str): Output ONNX model path.
         input_metadata (Dict[str, Any]): List of input metadata.
+        input_names (List[str]): List of model input names.
+        output_names (List[str]): List of model output names.
         batch_dim (Optional[int]): Batch dimension.
         target_device (str): Device to load TorchScript model on.
-        strict (bool): Enable or Disable strict flag for tracer used in TorchScript export.
+        dynamic_shapes (bool): Enable dynamic shapes.
+        verbose (bool): Enable verbose mode.
         navigator_workspace (Optional[str], optional): Model Navigator workspace path.
             When None use current workdir. Defaults to None.
         custom_args (Optional[Dict[str, Any]], optional): Passthrough parameters for torch.jit.trace
@@ -70,14 +80,56 @@ def export(
         dummy_input = (*dummy_input, {})
     *args, kwargs = dummy_input
 
+    loglevel = logging.DEBUG if verbose else logging.WARNING
+    export_options_kwargs = {}
+    export_options_kwargs["op_level_debug"] = verbose
+    export_options_kwargs["diagnostic_options"] = torch.onnx.DiagnosticOptions(verbosity_level=loglevel)
+    if dynamic_shapes:
+        export_options_kwargs["dynamic_shapes"] = True
+    export_options = torch.onnx.ExportOptions(**export_options_kwargs)
+
+    root_logger = logging.getLogger()
+    root_loglevel = root_logger.getEffectiveLevel()
+    root_logger.setLevel(loglevel)
+
     exported_model = torch.onnx.dynamo_export(
         model,
         *args,
         **custom_args,
         **kwargs,
+        export_options=export_options,
     )
 
     exported_model_path = pathlib.Path(exported_model_path)
     if not exported_model_path.is_absolute():
         exported_model_path = navigator_workspace / exported_model_path
     exported_model.save(exported_model_path.as_posix())
+
+    root_logger.setLevel(root_loglevel)
+
+    _modify_onnx_io_names(exported_model_path, input_names, output_names, exported_model_path)
+
+
+def _modify_onnx_io_names(model_path, new_input_names, new_output_names, output_path):
+    # Load the ONNX model
+    graph = gs.import_onnx(onnx.load(model_path))
+
+    # Check if the number of new input names matches the number of inputs in the graph
+    if len(new_input_names) != len(graph.inputs):
+        raise ValueError("Number of new input names must match the number of inputs in the ONNX graph.")
+
+    # Modify the input names
+    for i, input in enumerate(graph.inputs):
+        input.name = new_input_names[i]
+
+    # Check if the number of new output names matches the number of outputs in the graph
+    if len(new_output_names) != len(graph.outputs):
+        raise ValueError("Number of new output names must match the number of outputs in the ONNX graph.")
+
+    # Modify the output names
+    for i, output in enumerate(graph.outputs):
+        output.name = new_output_names[i]
+
+    # Save the modified model
+    graph.cleanup()
+    onnx.save(gs.export_onnx(graph), output_path)